// types/patient.ts
export interface Patient {
  firstName: string;
  middleName?: string;
  lastName: string;
  dateOfBirth: string;
  gender: string;
  phoneNumber: string;
  email: string;
  address: string;
  preferredLanguage: string;
  nationality: string;
  emergencyContact?: {
    name: string;
    relationship: string;
  };
  religion?: string;
}